updated derelict-gl3 to last minor version
from 1.0.16 to 1.0.18. This speeds up dub package resolving, because dub in the current release needs some time to do this automatically.
beside from that this commit does not have any effect because ~>1.0.16 and ~>1.0.18 both will be resolved to the highest available package which is 1.0.18. i checked all packages from the dub.selection.json.