FileDialog: Improve directory selection
If you select a directory but do not open it and then hit the "Select Directory" button to close the dialog the open path would be given as result instead of the selected directory.
This change breaks backwards compatibility insofar that the "path" property you would use to access the selected path in OpenDirectory dialogs is no longer the path the user chose, but instead the parent directory of the path the user chose.