Call correctCaretPos() after keyboard navigation
If the up/down arrow keys were used for keyboard navigation in a
multiline editbox, the position of the caret would become invalid
because it was not updated to fit in the new line. This caused
incorrect behaviour when entering text after such navigation
(ie replacing text when it should have appended).